feat: more work

+++ src/entities/token_accts.rs
@@ -5,7 +5,7 @@ table! {
         token_acct -> Varchar,
         mint_acct -> Varchar,
         owner_acct -> Varchar,
-        amount -> Float,
+        amount -> Double,
         updated_at -> Nullable<Timestamp>,
     }
 }
@@ -15,6 +15,6 @@ pub struct TokenAcct {
     pub token_acct: String,
     pub mint_acct: String,
     pub owner_acct: String,
-    pub amount: f32,
+    pub amount: f64,
     pub updated_at: Option<SystemTime>,
 }
